Main/Unattended: Detect OS/2 ISOs. [build fix]
svn:sync-xref-src-repo-rev: r149095

+++ src/VBox/Main/src-server/UnattendedImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: UnattendedImpl.cpp 93082 2021-12-26 01:27:48Z knut.osmundsen@oracle.com $ */
+/* $Id: UnattendedImpl.cpp 93083 2021-12-26 01:44:52Z knut.osmundsen@oracle.com $ */
 /** @file
  * Unattended class implementation
  */
@@ -1338,7 +1338,7 @@ HRESULT Unattended::i_innerDetectIsoOSOs2(RTVFS hVfsIso, DETECTBUFFER *pBuf, VBO
                 if (pszNewLine && RTStrStartsWith(pszNewLine + 1, "Built 20"))
                 {
                     size_t offRemove = 0;
-                    while (RT_C_IS_SPACE(pszNewLine[-1 - offRemove]))
+                    while (RT_C_IS_SPACE(pszNewLine[-1 - (ssize_t)offRemove]))
                         offRemove++;
                     if (offRemove > 0)
                     {
